# Improve performance of time panel