--- a/crates/viewer/re_time_panel/src/lib.rs
+++ b/crates/viewer/re_time_panel/src/lib.rs
@@ -6,6 +6,7 @@
 // TODO(#6330): remove unwrap()
 #![allow(clippy::unwrap_used)]
 
+mod chunk_statistics_store_subscriber;
 mod data_density_graph;
 mod paint_ticks;
 mod time_axis;