@emilk 2024-11-26
> Attaching the profiler doesn't tell us much since the profiler overhead drowns out everything else

If you are using the latest `puffin_viewer`, you should see a warning when that happens. If you go to the `Table` view you can see which profile scopes are too frequent. Remove them!