@Wumpf 2024-11-26
I did, that was in the _before_ state and was actually the first thing I committed :). Didn't bother though to restart my profiling for the PR description at an arbitrary commit and rather used latest main (which moved on quite a bit since I first removed those overhead inducing profile scopes).
I think I no longer got the warning (which I did get before indeed!), but there's still plenty of overhead that I should probably address!